refactor(entreprise): déplace RÈGLES_IDENTITÉ_ENTREPRISE dans le domaine et la consomme directement depuis la fonction de décision
