+++ site/source/domaine/Entreprise.ts
@@ -1,6 +1,7 @@
 import { CodeActivite } from '@/domaine/CodeActivite'
 import { CodeCatégorieJuridique } from '@/domaine/CodeCatégorieJuridique'
 import { Établissement } from '@/domaine/Établissement'
+import { DottedName } from '@/domaine/publicodes/DottedName'
 import { Siren } from '@/domaine/Siren'
 
 export interface Entreprise {
@@ -12,6 +13,18 @@ export interface Entreprise {
 	siège?: Établissement
 	établissement: Établissement
 }
+
+export const RÈGLES_IDENTITÉ_ENTREPRISE = [
+	'entreprise . date de création',
+	'entreprise . code catégorie juridique',
+	'entreprise . catégorie juridique',
+	'entreprise . SIREN',
+	'entreprise . nom',
+	'établissement . SIRET',
+] as const satisfies ReadonlyArray<DottedName>
+
+export type RègleIdentitéEntreprise =
+	(typeof RÈGLES_IDENTITÉ_ENTREPRISE)[number]
 
 const établissementEstLeSiège = (entreprise: Entreprise): boolean =>
 	!!entreprise.siège &&
